TINKERPOP-1878 Major refactoring of sparql-gremlin classes
Renamed the "compiler" to a "transpiler" since it's really transforming sparql to a language of a similar level of abstraction in Gremlin. Lots of code reformatting and dead code pruning.
